Moved Google Play features to their own package.
