+++ app/src/main/java/com/rohankhayech/choona/controller/play/ReviewController.kt
@@ -16,7 +16,7 @@
  * along with this program.  If not, see <https://www.gnu.org/licenses/>.
  */
 
-package com.rohankhayech.choona.controller
+package com.rohankhayech.choona.controller.play
 
 /**
  * Controller that handles launching Google Play review prompts.
